requirements.txt
dspy
numpy
pyyaml